Continuing to re-organize files. Practice module migrated, PDF module migrated. Working on session and summary
