--- src/containers/question/QuestionContainer.module.css
@@ -1,11 +1,0 @@
-
-.questionSize{
-  max-width: 70vw;
-  height: auto;
-}
-
-.questionAlign{
-  composes: flexColCenter from "@/src/App.module.css";
-  margin-right: 2rem;
-  margin-left: 4rem;
-}
