+++ src/features/practice/Practice.module.css
@@ -1,0 +1,3 @@
+.container {
+  composes: flexColCenter from "@/src/App.module.css";
+}
