+++ src/features/practice/components/Practice.questionImage.tsx
