+++ src/features/practice/containers/PracticeContainer.module.css
@@ -33,3 +33,14 @@
     background-color: lightblue;
   }
 }
+
+.questionSize{
+  max-width: 70vw;
+  height: auto;
+}
+
+.questionAlign{
+  composes: flexColCenter from "@/src/App.module.css";
+  margin-right: 2rem;
+  margin-left: 4rem;
+}
