+++ src/pages/home.tsx
@@ -1,6 +1,6 @@
 import HomeContainer from "containers/home/HomeContainer";
 
-export default function Home() {
+export default function HomePage() {
   return (
     <>
       <HomeContainer />
