+++ src/pages/practice.tsx
@@ -1,9 +1,9 @@
-import PracticeContainer from "containers/practice/PracticeContainer"
+import Practice from "@/src/features/practice/Practice.tsx"
 
 // The practice route will render the practice module.
-export default function Practice() {
+export default function PracticePage() {
 
   return (
-    <PracticeContainer />
+    <Practice />
   )
 }
